+++ client/src/pages/MailingList/events.ts
