+++ client/src/pages/MailingList/slides/Email.tsx
